Refactored edu courses
