+++ package.json
@@ -6,7 +6,7 @@
   "author": "Decentralize Justice",
   "scripts": {
     "serve": "vue-cli-service electron:serve",
-    "build": "npm-run-all buildNativeMod electron:generate-icons electron:build",
+    "build": "npm-run-all electron:generate-icons electron:build",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "vue-cli-service test:e2e",
     "lint": "vue-cli-service lint",
