+++ src/components/general/congrats.vue
@@ -1,13 +1,13 @@
 <template>
-  <v-row align="center">
-    <v-col cols='6' class="text-center" offset='3'>
-      <v-alert type="success">
+  <v-row align="center" justify='center'>
+    <v-col cols='12' class="text-center">
+      <v-alert type="success" style="margin:auto;width:50%;">
         Congratulations !!!
       </v-alert>
     </v-col>
-   <v-col cols='8' offset='2'>
+   <v-col cols='4'>
      <videoPlayer
-     v-bind:url="url"
+      v-bind:vidUrl="vidUrl"
      />
    </v-col>
    <v-col class="text-center" cols="12">
@@ -22,7 +22,7 @@
 </template>
 
 <script>
-import videoPlayer from '@/components/general/videoHashed.vue'
+import videoPlayer from '@/components/general/localVideo.vue'
 export default {
   name: 'congrats',
   components: {
@@ -39,11 +39,8 @@ export default {
     }
   },
   computed: {
-    url: function () {
-      const path = require('path')
-      const fileLocation = path.join(process.env.BASE_URL, this.vidUrl)
-      return fileLocation
-    }
+  },
+  async mounted () {
   }
 }
 </script>
