+++ src/components/general/hashCheckVid.vue
