--- src/components/general/incompleteIntro.vue
@@ -1,39 +1,0 @@
-<template>
-  <v-card class="text-xs-center no-gutters" style="">
-    <v-card-title class="headline justify-center">
-      Incomplete Intro
-    </v-card-title>
-        <v-divider></v-divider>
-      <v-divider/>
-      <v-card-actions>
-      <v-btn
-        color="orange"
-        text
-      >
-        <v-icon>mdi-help</v-icon>
-      </v-btn>
-      <v-btn
-        color="primary"
-        text
-      >
-        <v-icon>mdi-refresh</v-icon>
-      </v-btn>
-    </v-card-actions>
-  </v-card>
-</template>
-
-<script>
-export default {
-  components: {
-  },
-  data: () => ({
-  }),
-  methods: {
-  },
-  computed: {
-  }
-}
-</script>
-<style scoped>
-
-</style>
